+++ stratosphere/src/services/aws/events.rs
@@ -620,7 +620,7 @@ pub mod rule {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-events-rule-batcharrayproperties.html
     pub struct BatchArrayProperties_ {
-        pub size: Option<i64>,
+        pub size: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -684,7 +684,7 @@ pub mod rule {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-events-rule-batchretrystrategy.html
     pub struct BatchRetryStrategy_ {
-        pub attempts: Option<i64>,
+        pub attempts: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -709,9 +709,9 @@ pub mod rule {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-events-rule-capacityproviderstrategyitem.html
     pub struct CapacityProviderStrategyItem_ {
-        pub base: Option<i64>,
+        pub base: Option<i32>,
         pub capacity_provider: crate::value::ExpString,
-        pub weight: Option<i64>,
+        pub weight: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -774,7 +774,7 @@ pub mod rule {
         pub propagate_tags: Option<crate::value::ExpString>,
         pub reference_id: Option<crate::value::ExpString>,
         pub tag_list: Option<Vec<crate::Tag_>>,
-        pub task_count: Option<i64>,
+        pub task_count: Option<i32>,
         pub task_definition_arn: crate::value::ExpString,
     }
     #[doc(hidden)]
@@ -1102,8 +1102,8 @@ pub mod rule {
     }
     ///http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-events-rule-retrypolicy.html
     pub struct RetryPolicy_ {
-        pub maximum_event_age_in_seconds: Option<i64>,
-        pub maximum_retry_attempts: Option<i64>,
+        pub maximum_event_age_in_seconds: Option<i32>,
+        pub maximum_retry_attempts: Option<i32>,
     }
     #[doc(hidden)]
     #[macro_export]
@@ -1393,7 +1393,7 @@ pub struct ApiDestination_ {
     pub description: Option<crate::value::ExpString>,
     pub http_method: crate::value::ExpString,
     pub invocation_endpoint: crate::value::ExpString,
-    pub invocation_rate_limit_per_second: Option<i64>,
+    pub invocation_rate_limit_per_second: Option<i32>,
     pub name: Option<crate::value::ExpString>,
 }
 #[doc(hidden)]
@@ -1452,7 +1452,7 @@ pub struct Archive_ {
     pub description: Option<crate::value::ExpString>,
     pub event_pattern: Option<serde_json::Value>,
     pub kms_key_identifier: Option<crate::value::ExpString>,
-    pub retention_days: Option<i64>,
+    pub retention_days: Option<i32>,
     pub source_arn: crate::value::ExpString,
 }
 #[doc(hidden)]
